Split replay lifecycle and replay tests to respect file length budget
MobileShellComposite.swift grew past main's file-length budget with the
replay fix, and the new regression tests grew the liveness test files the
same way. Move the replay barrier / replay-request lifecycle into
MobileShellComposite+TerminalReplayLifecycle.swift (widening the replay
lifecycle storage to internal, following the +MacSwitchState precedent),
split the cold-attach barrier and replay-staleness tests into their own
files, and move the shared event-frame fixture builders into
MobileShellRenderGridEventFrameFixtures.swift. Every budget entry is back
at or below main's ceiling; no new tracked entries.
