+++ Packages/iOS/CmuxMobileShell/Tests/CmuxMobileShellTests/MobileShellRenderGridLivenessTests.swift
@@ -124,706 +124,6 @@ import Testing
 }
 
 @MainActor
-@Test func coldAttachReplayBarrierHoldsRenderGridDeltaUntilBaseApplies() async throws {
-    let clock = TestClock()
-    let router = LivenessHostRouter()
-    await router.setCapabilities(["events.v1", "terminal.render_grid.v1", "terminal.replay.v1"])
-    let box = TransportBox()
-    let store = try await makeConnectedStore(router: router, box: box, clock: clock)
-    defer {
-        Task { await router.releaseAllHeld() }
-    }
-
-    let collector = OutputCollector()
-    let replayCapabilityResolved = try await pollUntil {
-        store.supportedHostCapabilities.contains("terminal.replay.v1")
-    }
-    #expect(replayCapabilityResolved, "the host replay capability must be known before mounting")
-    await router.holdNextReplayResponses()
-    collector.mount(store: store, surfaceID: "live-terminal")
-    let coldReplayRequested = try await pollUntil { await router.count(of: "mobile.terminal.replay") >= 1 }
-    #expect(coldReplayRequested, "mounting a sink must request the cold replay")
-    let transport = try #require(box.get())
-
-    await transport.deliver(try renderGridEventFrame(
-        surfaceID: "live-terminal",
-        seq: 6,
-        text: "racing-delta",
-        columns: 16,
-        full: false
-    ))
-    let preBaseRendered = try await pollUntil(attempts: 60) {
-        collector.lines.contains { $0.contains("racing-delta") }
-    }
-    #expect(
-        preBaseRendered == false,
-        "a cold-attach render-grid delta must not paint into an empty local terminal before the authoritative replay base applies"
-    )
-
-    await router.enqueueReplayRenderGridFrames([
-        try MobileTerminalRenderGridFrame(
-            surfaceID: "live-terminal",
-            stateSeq: 5,
-            columns: 24,
-            rows: 4,
-            full: true,
-            rowSpans: [
-                .init(row: 0, column: 0, text: "authoritative-base"),
-            ]
-        ),
-    ])
-    await router.releaseAllHeld()
-
-    let baseDelivered = try await pollUntil {
-        collector.lines.contains { $0.contains("authoritative-base") }
-    }
-    #expect(
-        baseDelivered,
-        "the cold replay base must still apply even when newer live output raced it"
-    )
-    let followUpSettled = try await pollUntil { await router.replayResponsesServed() >= 2 }
-    #expect(followUpSettled, "dropped racing output should trigger and settle one catch-up replay")
-
-    await transport.deliver(try renderGridEventFrame(
-        surfaceID: "live-terminal",
-        seq: 7,
-        text: "post-settle",
-        columns: 16
-    ))
-    let postSettleDelivered = try await pollUntil {
-        collector.lines.contains { $0.contains("post-settle") }
-    }
-    #expect(postSettleDelivered, "live render-grid output must resume after the cold barrier settles")
-    collector.unmount()
-}
-
-@MainActor
-@Test func coldAttachReplayBarrierHoldsHybridRawBytesUntilBaseApplies() async throws {
-    let clock = TestClock()
-    let router = LivenessHostRouter()
-    let box = TransportBox()
-    let store = try await makeConnectedStore(router: router, box: box, clock: clock)
-    defer {
-        Task { await router.releaseAllHeld() }
-    }
-
-    let collector = OutputCollector()
-    let replayCapabilityResolved = try await pollUntil {
-        store.supportedHostCapabilities.contains("terminal.replay.v1")
-    }
-    #expect(replayCapabilityResolved, "the host replay capability must be known before mounting")
-    await router.holdNextReplayResponses()
-    collector.mount(store: store, surfaceID: "live-terminal")
-    let coldReplayRequested = try await pollUntil { await router.count(of: "mobile.terminal.replay") >= 1 }
-    #expect(coldReplayRequested, "mounting a sink must request the cold replay")
-    let transport = try #require(box.get())
-
-    await transport.deliver(try terminalBytesEventFrame(
-        surfaceID: "live-terminal",
-        seq: 6,
-        text: "racing-raw"
-    ))
-    let preBaseRendered = try await pollUntil(attempts: 60) {
-        collector.lines.contains { $0.contains("racing-raw") }
-    }
-    #expect(
-        preBaseRendered == false,
-        "cold-attach raw bytes must not paint into an empty local terminal before the authoritative replay base applies"
-    )
-
-    await router.enqueueReplayRenderGridFrames([
-        try MobileTerminalRenderGridFrame(
-            surfaceID: "live-terminal",
-            stateSeq: 5,
-            columns: 24,
-            rows: 4,
-            full: true,
-            rowSpans: [
-                .init(row: 0, column: 0, text: "authoritative-base"),
-            ]
-        ),
-    ])
-    await router.releaseAllHeld()
-
-    let baseDelivered = try await pollUntil {
-        collector.lines.contains { $0.contains("authoritative-base") }
-    }
-    #expect(
-        baseDelivered,
-        "the cold replay base must still apply even when newer raw bytes raced it"
-    )
-    let followUpSettled = try await pollUntil { await router.replayResponsesServed() >= 2 }
-    #expect(followUpSettled, "dropped racing output should trigger and settle one catch-up replay")
-
-    await transport.deliver(try terminalBytesEventFrame(
-        surfaceID: "live-terminal",
-        seq: 5,
-        text: "post-raw"
-    ))
-    let postSettleDelivered = try await pollUntil {
-        collector.lines.contains { $0.contains("post-raw") }
-    }
-    #expect(postSettleDelivered, "live raw output must resume after the cold barrier settles")
-    collector.unmount()
-}
-
-@MainActor
-@Test func coldAttachReplayMountedBeforeConnectionUpgradesToBarrierWhenCapabilitiesResolve() async throws {
-    let clock = TestClock()
-    let router = LivenessHostRouter()
-    await router.setCapabilities(["events.v1", "terminal.render_grid.v1", "terminal.replay.v1"])
-    await router.holdNextReplayResponses()
-    let box = TransportBox()
-    let runtime = LivenessTestRuntime(
-        transportFactory: LivenessTransportFactory(router: router, box: box),
-        now: { clock.now }
-    )
-    let store = MobileShellComposite.preview(runtime: runtime)
-    store.signIn()
-    defer {
-        Task { await router.releaseAllHeld() }
-    }
-
-    let collector = OutputCollector()
-    collector.mount(store: store, surfaceID: "live-terminal")
-    let replayBeforeConnection = try await pollUntil(attempts: 60) {
-        await router.count(of: "mobile.terminal.replay") > 0
-    }
-    #expect(
-        replayBeforeConnection == false,
-        "mounting before a remote client exists must not send an inert replay request"
-    )
-
-    let connected = await store.connectPairingURL(try attachURL(for: makeTicket(clock: clock)))
-    #expect(connected, "scripted connect must succeed")
-    let capabilitiesResolved = try await pollUntil {
-        store.supportedHostCapabilities.contains("terminal.replay.v1")
-    }
-    #expect(capabilitiesResolved, "host capabilities must resolve after connecting")
-    let coldReplayRequested = try await pollUntil {
-        await router.count(of: "mobile.terminal.replay") >= 1
-    }
-    #expect(
-        coldReplayRequested,
-        "a sink mounted before connection must be upgraded to a barriered replay once replay capability is known"
-    )
-    let transport = try #require(box.get())
-
-    await transport.deliver(try renderGridEventFrame(
-        surfaceID: "live-terminal",
-        seq: 6,
-        text: "early-delta",
-        columns: 16,
-        full: false
-    ))
-    let preBaseRendered = try await pollUntil(attempts: 60) {
-        collector.lines.contains { $0.contains("early-delta") }
-    }
-    #expect(
-        preBaseRendered == false,
-        "the deferred cold replay upgrade must hold racing deltas until the base replay applies"
-    )
-
-    await router.enqueueReplayRenderGridFrames([
-        try MobileTerminalRenderGridFrame(
-            surfaceID: "live-terminal",
-            stateSeq: 5,
-            columns: 24,
-            rows: 4,
-            full: true,
-            rowSpans: [
-                .init(row: 0, column: 0, text: "deferred-base"),
-            ]
-        ),
-    ])
-    await router.releaseAllHeld()
-
-    let baseDelivered = try await pollUntil {
-        collector.lines.contains { $0.contains("deferred-base") }
-    }
-    #expect(
-        baseDelivered,
-        "the deferred cold replay base must apply after capabilities resolve"
-    )
-    let followUpSettled = try await pollUntil { await router.replayResponsesServed() >= 2 }
-    #expect(followUpSettled, "dropped pre-base output should still trigger and settle one catch-up replay")
-    collector.unmount()
-}
-
-@MainActor
-@Test func renderGridReplayAtSameSeqDoesNotOverwriteNewerLiveGrid() async throws {
-    let clock = TestClock()
-    let router = LivenessHostRouter()
-    await router.setCapabilities(["events.v1", "terminal.render_grid.v1", "terminal.replay.v1"])
-    let box = TransportBox()
-    let store = try await makeConnectedStore(router: router, box: box, clock: clock)
-    defer {
-        Task { await router.releaseAllHeld() }
-    }
-
-    let collector = OutputCollector()
-    collector.mount(store: store, surfaceID: "live-terminal")
-    let coldReplayRequested = try await pollUntil { await router.count(of: "mobile.terminal.replay") >= 1 }
-    #expect(coldReplayRequested, "mounting a sink must request the cold replay")
-    try await waitForReplayResponsesServed(
-        1,
-        router: router,
-        "the cold replay response must settle before arming the held non-cold replay"
-    )
-    let transport = try #require(box.get())
-
-    await router.holdNextReplayResponses()
-    let replayCountBeforeHeldReplay = await router.count(of: "mobile.terminal.replay")
-    store.requestTerminalReplay(surfaceID: "live-terminal")
-    let heldReplayRequested = try await pollUntil {
-        await router.count(of: "mobile.terminal.replay") > replayCountBeforeHeldReplay
-    }
-    #expect(heldReplayRequested, "the non-cold replay must be requested before the live grid paints")
-
-    await transport.deliver(try renderGridEventFrame(
-        surfaceID: "live-terminal",
-        seq: 5,
-        text: "fresh-wide-grid",
-        columns: 16
-    ))
-    let freshDelivered = try await pollUntil {
-        collector.lines.contains { $0.contains("fresh-wide-grid") }
-    }
-    #expect(freshDelivered, "the live render-grid frame must paint before the held replay resolves")
-
-    await router.enqueueReplayRenderGridFrames([
-        try MobileTerminalRenderGridFrame(
-            surfaceID: "live-terminal",
-            stateSeq: 5,
-            columns: 4,
-            rows: 4,
-            full: true,
-            rowSpans: [
-                .init(row: 0, column: 0, text: "old!"),
-            ]
-        ),
-    ])
-    await router.releaseAllHeld()
-
-    let staleDelivered = try await pollUntil(attempts: 60) {
-        collector.lines.contains { $0.contains("old!") }
-    }
-    #expect(
-        staleDelivered == false,
-        "a replay captured at an older grid width must not overwrite an already-delivered live frame at the same state sequence"
-    )
-    collector.unmount()
-}
-
-@MainActor
-@Test func hybridPrimaryFullGridAllowsSameSeqReplayBeforeRawBytesCatchUp() async throws {
-    let clock = TestClock()
-    let router = LivenessHostRouter()
-    let box = TransportBox()
-    let store = try await makeConnectedStore(router: router, box: box, clock: clock)
-    defer {
-        Task { await router.releaseAllHeld() }
-    }
-
-    let collector = OutputCollector()
-    collector.mount(store: store, surfaceID: "live-terminal")
-    let coldReplayRequested = try await pollUntil { await router.count(of: "mobile.terminal.replay") >= 1 }
-    #expect(coldReplayRequested, "mounting a sink must request the cold replay")
-    try await waitForReplayResponsesServed(
-        1,
-        router: router,
-        "the cold replay response must settle before arming the held non-cold replay"
-    )
-    let transport = try #require(box.get())
-
-    await router.holdNextReplayResponses()
-    let replayCountBeforeHeldReplay = await router.count(of: "mobile.terminal.replay")
-    store.requestTerminalReplay(surfaceID: "live-terminal")
-    let heldReplayRequested = try await pollUntil {
-        await router.count(of: "mobile.terminal.replay") > replayCountBeforeHeldReplay
-    }
-    #expect(heldReplayRequested, "the non-cold replay must be requested before the advisory grid")
-
-    await transport.deliver(try renderGridEventFrame(
-        surfaceID: "live-terminal",
-        seq: 5,
-        text: "fresh-wide-grid",
-        columns: 16
-    ))
-    let advisoryProcessed = try await pollUntil {
-        collector.viewportPolicies.last == .natural
-    }
-    #expect(advisoryProcessed, "primary render-grid events are advisory in default hybrid mode")
-    #expect(
-        collector.lines.contains { $0.contains("fresh-wide-grid") } == false,
-        "the hybrid advisory path must not advance raw-byte delivery"
-    )
-
-    await router.enqueueReplayRenderGridFrames([
-        try MobileTerminalRenderGridFrame(
-            surfaceID: "live-terminal",
-            stateSeq: 5,
-            columns: 4,
-            rows: 4,
-            full: true,
-            rowSpans: [
-                .init(row: 0, column: 0, text: "old!"),
-            ]
-        ),
-    ])
-    await router.releaseAllHeld()
-
-    let replayDelivered = try await pollUntil {
-        collector.lines.contains { $0.contains("old!") }
-    }
-    #expect(
-        replayDelivered,
-        "an advisory primary full grid does not paint terminal content, so the same-sequence replay must seed the local terminal until raw bytes catch up"
-    )
-    collector.unmount()
-}
-
-@MainActor
-@Test func hybridPrimaryFullGridStillSuppressesSameSeqReplayAfterRawBytesCatchUp() async throws {
-    let clock = TestClock()
-    let router = LivenessHostRouter()
-    let box = TransportBox()
-    let store = try await makeConnectedStore(router: router, box: box, clock: clock)
-    defer {
-        Task { await router.releaseAllHeld() }
-    }
-
-    let collector = OutputCollector()
-    collector.mount(store: store, surfaceID: "live-terminal")
-    let coldReplayRequested = try await pollUntil { await router.count(of: "mobile.terminal.replay") >= 1 }
-    #expect(coldReplayRequested, "mounting a sink must request the cold replay")
-    try await waitForReplayResponsesServed(
-        1,
-        router: router,
-        "the cold replay response must settle before arming the held non-cold replay"
-    )
-    let transport = try #require(box.get())
-
-    await router.holdNextReplayResponses()
-    let replayCountBeforeHeldReplay = await router.count(of: "mobile.terminal.replay")
-    store.requestTerminalReplay(surfaceID: "live-terminal")
-    let heldReplayRequested = try await pollUntil {
-        await router.count(of: "mobile.terminal.replay") > replayCountBeforeHeldReplay
-    }
-    #expect(heldReplayRequested, "the non-cold replay must be requested before raw bytes catch up")
-
-    await transport.deliver(try renderGridEventFrame(
-        surfaceID: "live-terminal",
-        seq: 5,
-        text: "fresh-wide-grid",
-        columns: 16
-    ))
-    let advisoryProcessed = try await pollUntil {
-        collector.viewportPolicies.last == .natural
-    }
-    #expect(advisoryProcessed, "primary render-grid events are advisory in default hybrid mode")
-
-    await transport.deliver(try terminalBytesEventFrame(surfaceID: "live-terminal", seq: 0, text: "raw5!"))
-    let rawDelivered = try await pollUntil { collector.lines.contains { $0.contains("raw5!") } }
-    #expect(rawDelivered, "same-sequence raw bytes must still paint in hybrid primary mode")
-
-    await router.enqueueReplayRenderGridFrames([
-        try MobileTerminalRenderGridFrame(
-            surfaceID: "live-terminal",
-            stateSeq: 5,
-            columns: 4,
-            rows: 4,
-            full: true,
-            rowSpans: [
-                .init(row: 0, column: 0, text: "old!"),
-            ]
-        ),
-    ])
-    await router.releaseAllHeld()
-
-    let staleDelivered = try await pollUntil(attempts: 60) {
-        collector.lines.contains { $0.contains("old!") }
-    }
-    #expect(
-        staleDelivered == false,
-        "same-sequence raw bytes must not clear the advisory full-grid freshness marker before a held replay resolves"
-    )
-    collector.unmount()
-}
-
-@MainActor
-@Test func hybridPrimaryNewerFullGridSuppressesOlderStaleReplay() async throws {
-    let clock = TestClock()
-    let router = LivenessHostRouter()
-    let box = TransportBox()
-    let store = try await makeConnectedStore(router: router, box: box, clock: clock)
-    defer {
-        Task { await router.releaseAllHeld() }
-    }
-
-    let collector = OutputCollector()
-    collector.mount(store: store, surfaceID: "live-terminal")
-    let coldReplayRequested = try await pollUntil { await router.count(of: "mobile.terminal.replay") >= 1 }
-    #expect(coldReplayRequested, "mounting a sink must request the cold replay")
-    try await waitForReplayResponsesServed(
-        1,
-        router: router,
-        "the cold replay response must settle before arming the held non-cold replay"
-    )
-    let transport = try #require(box.get())
-
-    await router.holdNextReplayResponses()
-    let replayCountBeforeHeldReplay = await router.count(of: "mobile.terminal.replay")
-    store.requestTerminalReplay(surfaceID: "live-terminal")
-    let heldReplayRequested = try await pollUntil {
-        await router.count(of: "mobile.terminal.replay") > replayCountBeforeHeldReplay
-    }
-    #expect(heldReplayRequested, "the non-cold replay must be requested before raw bytes cover the older seq")
-
-    await transport.deliver(try renderGridEventFrame(
-        surfaceID: "live-terminal",
-        seq: 6,
-        text: "newer-grid",
-        columns: 16
-    ))
-    let advisoryProcessed = try await pollUntil {
-        collector.viewportPolicies.last == .natural
-    }
-    #expect(advisoryProcessed, "the newer primary full grid is advisory in default hybrid mode")
-
-    await transport.deliver(try terminalBytesEventFrame(surfaceID: "live-terminal", seq: 0, text: "raw5!"))
-    let rawDelivered = try await pollUntil { collector.lines.contains { $0.contains("raw5!") } }
-    #expect(rawDelivered, "raw bytes covering the older replay seq must paint before the held replay resolves")
-
-    await router.enqueueReplayRenderGridFrames([
-        try MobileTerminalRenderGridFrame(
-            surfaceID: "live-terminal",
-            stateSeq: 5,
-            columns: 16,
-            rows: 4,
-            full: true,
-            rowSpans: [
-                .init(row: 0, column: 0, text: "older-replay"),
-            ]
-        ),
-    ])
-    await router.releaseAllHeld()
-
-    let staleDelivered = try await pollUntil(attempts: 60) {
-        collector.lines.contains { $0.contains("older-replay") }
-    }
-    #expect(
-        staleDelivered == false,
-        "a newer hybrid primary full-grid observation plus raw byte coverage must stale an older held replay"
-    )
-    collector.unmount()
-}
-
-@MainActor
-@Test func hybridPrimaryFullGridDuringReplayBarrierDoesNotSuppressBarrierReplay() async throws {
-    let clock = TestClock()
-    let router = LivenessHostRouter()
-    let box = TransportBox()
-    let store = try await makeConnectedStore(router: router, box: box, clock: clock)
-    defer {
-        Task { await router.releaseAllHeld() }
-    }
-
-    let collector = OutputCollector()
-    collector.mount(store: store, surfaceID: "live-terminal")
-    let sawMountReplay = try await pollUntil { await router.count(of: "mobile.terminal.replay") >= 1 }
-    #expect(sawMountReplay, "mounting a sink must request the cold replay")
-    let mountReplayCompleted = try await pollUntil { await router.replayResponsesServed() >= 1 }
-    #expect(mountReplayCompleted, "the cold replay response must complete before arming the barrier hold")
-    let transport = try #require(box.get())
-
-    await router.holdNextReplayResponses()
-    let replayCountBeforeBarrier = await router.count(of: "mobile.terminal.replay")
-    store.terminalOutputNeedsReplay(surfaceID: "live-terminal")
-    let barrierReplayRequested = try await pollUntil {
-        await router.count(of: "mobile.terminal.replay") > replayCountBeforeBarrier
-    }
-    #expect(barrierReplayRequested, "manual replay must create a replay barrier request")
-
-    let viewportPolicyCountBeforeAdvisory = collector.viewportPolicies.count
-    await transport.deliver(try renderGridEventFrame(
-        surfaceID: "live-terminal",
-        seq: 5,
-        text: "barrier-advisory",
-        columns: 16
-    ))
-    let advisoryDeliveredDuringBarrier = try await pollUntil(attempts: 60) {
-        collector.viewportPolicies.count > viewportPolicyCountBeforeAdvisory
-    }
-    #expect(
-        advisoryDeliveredDuringBarrier == false,
-        "advisory output is dropped while the replay barrier waits for the authoritative replay"
-    )
-
-    await router.enqueueReplayRenderGridFrames([
-        try MobileTerminalRenderGridFrame(
-            surfaceID: "live-terminal",
-            stateSeq: 5,
-            columns: 16,
-            rows: 4,
-            full: true,
-            rowSpans: [
-                .init(row: 0, column: 0, text: "barrier-replay"),
-            ]
-        ),
-    ])
-    await router.releaseAllHeld()
-
-    let replayDelivered = try await pollUntil {
-        collector.lines.contains { $0.contains("barrier-replay") }
-    }
-    #expect(
-        replayDelivered,
-        "a same-seq barrier replay must still apply after an advisory full grid that was dropped by the barrier"
-    )
-    collector.unmount()
-}
-
-@MainActor
-@Test func renderGridReplayAtSameSeqStillAppliesAfterPartialLiveDelta() async throws {
-    let clock = TestClock()
-    let router = LivenessHostRouter()
-    await router.setCapabilities(["events.v1", "terminal.render_grid.v1", "terminal.replay.v1"])
-    let box = TransportBox()
-    let store = try await makeConnectedStore(router: router, box: box, clock: clock)
-    defer {
-        Task { await router.releaseAllHeld() }
-    }
-
-    let collector = OutputCollector()
-    collector.mount(store: store, surfaceID: "live-terminal")
-    let coldReplayRequested = try await pollUntil { await router.count(of: "mobile.terminal.replay") >= 1 }
-    #expect(coldReplayRequested, "mounting a sink must request the cold replay")
-    try await waitForReplayResponsesServed(
-        1,
-        router: router,
-        "the cold replay response must settle before arming the held non-cold replay"
-    )
-    let transport = try #require(box.get())
-
-    await router.holdNextReplayResponses()
-    let replayCountBeforeHeldReplay = await router.count(of: "mobile.terminal.replay")
-    store.requestTerminalReplay(surfaceID: "live-terminal")
-    let heldReplayRequested = try await pollUntil {
-        await router.count(of: "mobile.terminal.replay") > replayCountBeforeHeldReplay
-    }
-    #expect(heldReplayRequested, "the non-cold replay must be requested before the partial delta")
-
-    await transport.deliver(try renderGridEventFrame(
-        surfaceID: "live-terminal",
-        seq: 5,
-        text: "partial-live-delta",
-        columns: 24,
-        full: false
-    ))
-    let partialDelivered = try await pollUntil {
-        collector.lines.contains { $0.contains("partial-live-delta") }
-    }
-    #expect(partialDelivered, "a live partial render-grid delta may arrive before the held replay")
-
-    await router.enqueueReplayRenderGridFrames([
-        try MobileTerminalRenderGridFrame(
-            surfaceID: "live-terminal",
-            stateSeq: 5,
-            columns: 24,
-            rows: 4,
-            full: true,
-            rowSpans: [
-                .init(row: 0, column: 0, text: "authoritative-snapshot"),
-            ]
-        ),
-    ])
-    await router.releaseAllHeld()
-
-    let replayDelivered = try await pollUntil {
-        collector.lines.contains { $0.contains("authoritative-snapshot") }
-    }
-    #expect(
-        replayDelivered,
-        "a same-sequence replay is still required when the only delivered live output was a partial delta"
-    )
-    collector.unmount()
-}
-
-@MainActor
-@Test func primaryRenderGridEventDoesNotPreemptRawBytes() async throws {
-    let clock = TestClock()
-    let router = LivenessHostRouter()
-    let box = TransportBox()
-    let store = try await makeConnectedStore(router: router, box: box, clock: clock)
-
-    let collector = OutputCollector()
-    collector.mount(store: store, surfaceID: "live-terminal")
-    let sawReplay = try await pollUntil { await router.count(of: "mobile.terminal.replay") >= 1 }
-    #expect(sawReplay, "mounting a sink must arm the cold-attach replay")
-    try await waitForReplayResponsesServed(
-        1,
-        router: router,
-        "the cold replay response must settle before testing advisory primary render-grid"
-    )
-    let transport = try #require(box.get())
-
-    await transport.deliver(try renderGridEventFrame(surfaceID: "live-terminal", seq: 3, text: "grid"))
-    let policyDelivered = try await pollUntil { collector.viewportPolicies.last == .natural }
-    #expect(policyDelivered, "advisory primary render-grid events must still deliver their viewport policy")
-    #expect(
-        collector.lines.contains { $0.contains("grid") } == false,
-        "primary render-grid events are advisory in hybrid mode; raw bytes own full-height primary rendering"
-    )
-
-    await transport.deliver(try terminalBytesEventFrame(surfaceID: "live-terminal", seq: 0, text: "raw"))
-    let rawDelivered = try await pollUntil { collector.lines.contains { $0.contains("raw") } }
-    #expect(rawDelivered, "advisory primary render-grid must not advance delivered seq and starve overlapping raw bytes")
-    #expect(
-        collector.lines.contains { $0.contains("grid") } == false,
-        "primary render-grid events are advisory in hybrid mode; raw bytes own full-height primary rendering"
-    )
-    #expect(collector.viewportPolicies.last == .natural)
-    collector.unmount()
-}
-
-@MainActor
-@Test func hybridPrimaryInputBehindRequestsReplayInsteadOfWaitingOnAdvisoryRenderGrid() async throws {
-    let clock = TestClock()
-    let router = LivenessHostRouter()
-    let box = TransportBox()
-    let store = try await makeConnectedStore(router: router, box: box, clock: clock)
-
-    let collector = OutputCollector()
-    collector.mount(store: store, surfaceID: "live-terminal")
-    let sawReplay = try await pollUntil { await router.count(of: "mobile.terminal.replay") >= 1 }
-    #expect(sawReplay, "mounting a sink must arm the cold-attach replay")
-    try await waitForReplayResponsesServed(
-        1,
-        router: router,
-        "the cold replay response must settle before testing input recovery"
-    )
-    let replayCountAfterMount = await router.count(of: "mobile.terminal.replay")
-    let transport = try #require(box.get())
-
-    await transport.deliver(try terminalBytesEventFrame(surfaceID: "live-terminal", seq: 0, text: "raw"))
-    let rawDelivered = try await pollUntil { collector.lines.contains { $0.contains("raw") } }
-    #expect(rawDelivered)
-
-    await store.submitTerminalRawInput(Data("x".utf8), surfaceID: "live-terminal")
-    let inputSent = try await pollUntil { await router.count(of: "terminal.input") >= 1 }
-    #expect(inputSent)
-    let replayRequested = try await pollUntil {
-        await router.count(of: "mobile.terminal.replay") > replayCountAfterMount
-    }
-    #expect(
-        replayRequested,
-        "hybrid primary output is advanced by terminal.bytes, so input recovery must request replay instead of waiting on advisory render-grid frames"
-    )
-    collector.unmount()
-}
-
-@MainActor
 @Test func alternateRenderGridPinsGridAndSuppressesRawBytes() async throws {
     let clock = TestClock()
     let router = LivenessHostRouter()
@@ -1196,62 +496,6 @@ import Testing
     await transport.deliver(event)
     let delivered = try await pollUntil { collector.lines.contains { $0.contains("repaired") } }
     #expect(delivered, "the original stream must still be consumed after the repair")
-    collector.unmount()
-}
-
-@MainActor
-@Test func livenessRepairDeliversSameSeqReplayAfterExistingFullGrid() async throws {
-    let clock = TestClock()
-    let router = LivenessHostRouter()
-    await router.setCapabilities(["events.v1", "terminal.render_grid.v1", "terminal.replay.v1"])
-    let box = TransportBox()
-    let store = try await makeConnectedStore(router: router, box: box, clock: clock)
-
-    let sawSubscribe = try await pollUntil { await router.count(of: "mobile.events.subscribe") >= 1 }
-    #expect(sawSubscribe, "listener must establish the push subscription")
-
-    let collector = OutputCollector()
-    collector.mount(store: store, surfaceID: "live-terminal")
-    let sawMountReplay = try await pollUntil { await router.count(of: "mobile.terminal.replay") >= 1 }
-    #expect(sawMountReplay, "mounting a sink arms the cold-attach replay")
-    let mountReplayCompleted = try await pollUntil { await router.replayResponsesServed() >= 1 }
-    #expect(mountReplayCompleted, "the cold replay response must complete before scripting the repair replay")
-    let transport = try #require(box.get())
-
-    await transport.deliver(try renderGridEventFrame(surfaceID: "live-terminal", seq: 5, text: "stale-grid"))
-    let staleGridDelivered = try await pollUntil {
-        collector.lines.contains { $0.contains("stale-grid") }
-    }
-    #expect(staleGridDelivered, "the pre-gap full render grid must establish the local delivered seq")
-
-    await router.dropSubscription()
-    await router.enqueueReplayRenderGridFrames([
-        try MobileTerminalRenderGridFrame(
-            surfaceID: "live-terminal",
-            stateSeq: 5,
-            columns: 16,
-            rows: 4,
-            full: true,
-            rowSpans: [
-                .init(row: 0, column: 0, text: "fresh-grid"),
-            ]
-        ),
-    ])
-    let replayCountBeforeRepair = await router.count(of: "mobile.terminal.replay")
-    clock.advance(by: 10)
-    store.debugRunRenderGridLivenessCheckForTesting()
-
-    let replayRequested = try await pollUntil {
-        await router.count(of: "mobile.terminal.replay") > replayCountBeforeRepair
-    }
-    #expect(replayRequested, "repairing a lost subscription must request a catch-up replay")
-    let freshGridDelivered = try await pollUntil {
-        collector.lines.contains { $0.contains("fresh-grid") }
-    }
-    #expect(
-        freshGridDelivered,
-        "a same-sequence recovery replay requested after an existing full grid must still repaint"
-    )
     collector.unmount()
 }
